scripts: awk script to get total replay time

+++ scripts/replay_collected_inputs.sh
@@ -11,17 +11,21 @@ fi
 replay_inputs() {
   RUNDIR=${1}
   TARGET=$(echo -n ${RUNDIR} | sed "s/RUNDIR-${FUZZING_ENGINE}-//g")
+  INPUT_DIR="${RUNDIR}/collect-seeds/inputs"
   REPLAY_LOG="${TARGET}-${FUZZING_ENGINE}-replay.log"
 
-  echo "Replaying ${TARGET}"
+  [[ ! -d ${INPUT_DIR} ]] && return
 
   rm -f ${REPLAY_LOG}
 
   for I in 0 25001 50001 75001 100001 125001 150001 175001 200001 225001 250001 275001 300001 325001 350001 375001 400001 425001 450001 475001; do
-    INPUTS=$(find "${RUNDIR}/collect-seeds/inputs" -type f | sort | tail -n +$I | head -25000)
+    INPUTS=$(find ${INPUT_DIR} -type f | sort | tail -n +$I | head -25000)
     perf stat -r3 --output ${REPLAY_LOG} --append  \
       "${RUNDIR}/${TARGET}-${FUZZING_ENGINE}" ${INPUTS} > /dev/null 2>&1
   done
+
+  TOTAL_TIME=$(awk -F ' ' '$0 ~ /task-clock/ {TOTAL_TIME += $1} END {printf "%f", TOTAL_TIME}' ${REPLAY_LOG})
+  echo "${TARGET} replay time = ${TOTAL_TIME} ms"
 }
 
 export -f replay_inputs
